refactor: code improvements for table component
