+++ src/components/overview/panel.tsx
@@ -17,21 +17,6 @@ export interface PanelProps {
   onOpen: () => void;
 }
 
-const tableColumns = ["Name", "Site", "Description", "Created"];
-
-const people = [
-  {
-    name: "Jane Cooper",
-    title: "Regional Paradigm Technician",
-    department: "Optimization",
-    role: "Admin",
-    email: "jane.cooper@example.com",
-    image:
-      "https://images.unsplash.com/photo-1494790108377-be9c29b29330?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=4&w=256&h=256&q=60",
-  },
-  // More people...
-];
-
 export default function Panel({ userId, onOpen }: PanelProps) {
   const { data } = useSWR<SiteData[]>(`/api/sites/${userId}`, fetcher);
